+++ packages/terra-props-table/src/PropsTable.jsx
@@ -70,7 +70,7 @@ const PropsTable = ({ componentName, src, ...customProps }) => {
    */
   const componentProps = componentMetaData.props;
 
-  const tableEntryClass = cx('prop-table-element');
+  const tableRowClass = cx('prop-table-row');
 
   return (
     <div dir="ltr" className="markdown-body">
@@ -91,16 +91,16 @@ const PropsTable = ({ componentName, src, ...customProps }) => {
             const type = determineType(prop.type);
 
             return (
-              <tr key={key} style={{ fontSize: '90%' }}>
-                <td className={tableEntryClass} style={{ fontWeight: 'bold' }}>{key}</td>
-                <td className={tableEntryClass}>{(prop.type ? type : '')}</td>
+              <tr className={tableRowClass} key={key} style={{ fontSize: '90%' }}>
+                <td style={{ fontWeight: 'bold' }}>{key}</td>
+                <td>{(prop.type ? type : '')}</td>
                 {(prop.required ?
-                  <td className={tableEntryClass} style={{ color: '#d53700' }}>required</td>
-              : <td className={tableEntryClass} style={{ color: '#444' }}>optional</td>)}
+                  <td style={{ color: '#d53700' }}>required</td>
+              : <td style={{ color: '#444' }}>optional</td>)}
                 {(prop.defaultValue ?
-                  <td className={tableEntryClass} style={{ fontWeight: 'bold' }}>{prop.defaultValue.value}</td>
-              : <td className={tableEntryClass} style={{ color: '#444' }}>none</td>)}
-                <td className={tableEntryClass}><Markdown src={prop.description} /></td>
+                  <td style={{ fontWeight: 'bold' }}>{prop.defaultValue.value}</td>
+              : <td style={{ color: '#444' }}>none</td>)}
+                <td><Markdown src={prop.description} /></td>
               </tr>
             );
           })}